Fix bug in sort-styles when props included shorthands
If the style properties had a 'isEitherShorthand' case, it would
return out of the whole checkIsSorted function instead of
continuing the for loop. This caused some properties to not be
reported as unsorted.
It would be sufficient to replace `return` with `continue`, but
there was a ESLint rule disallowing the usage of continue